TASK 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -3,22 +3,22 @@
 /**
  * array_iterator - executes a function
  * given as a parameter on each element
- *                  of an array.
- * @array: the array to iterate through.
- * @size: the size of the array.
- * @action: a pointer to the function to use.
+ * @array: pointer to the array
+ * @size: size of the array
+ * @action: pointer to the function to use
  *
- * Return: Nothing.
+ * Return: Nothing
  */
+
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-    size_t i;
+	unsigned int i;
 
-    if (array != NULL && action != NULL)
-    {
-        for (i = 0; i < size; i++)
-        {
-            action(array[i]);
-        }
-    }
+	if (array != NULL && action != NULL)
+	{
+		for (i = 0; i < size; i++)
+		{
+			action(array[i]);
+		}
+	}
 }
